+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Fixed
 
 - Missing ResizeObserver throwing an error on some platforms
+- Profile link being visible on home page when user was not authenticated
 
 # 1.5.0